software/client/pyproject.toml
[project]
name = "dbay"
version = "0.3.0"
description = "Unified dual-mode client (HTTP GUI + direct UDP/Serial) for Device Bay VME racks"

readme = "README.md"
requires-python = ">=3.8"
dependencies = ["pydantic>=2.10.6", "pyserial>=3.5", "requests>=2.32.3"]
authors = [{ name = "Andrew Mueller", email = "andrewstermueller@gmail.com" }]
license = { text = "MIT" }
keywords = ["dbay", "client", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
urls = { Homepage = "https://github.com/bkorzh/dbay", Documentation = "https://bkorzh.github.io/dbay/" }

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dbay", "dbay.modules", "dbay.addons"]

[dependency-groups]
dev = ["pytest>=8.3.5"]
